Fixed functionality
I went through the same steps as the ones in the course lessons. I changed the port on the express server and installed cors to allow for interaction with external pages.

Now the handleFunction function is working and the express server and development servers are communicating. I should now be in a position to be able to implement the api call required in the project.