MiniCssExtractPlugin
Added this plugin. There are two more plugins to add that will minify the styles. I'm committing before as during the course the next two caused an error. Having this as a point to fall back on if necessary gives me peace of mind.

(also I updated the loader rule in wenpack.prod.js, correcting it from the last commit)