API response
I'm now getting a response from the api call and I've started to add styling for displaying info to the user.

In the next commit I'll continue with the styling and displaying info from the api call.

NOTE: The api key is currently hard-coded in formhandler.js. After I get the display to where I want it, I'll work on retrieving the key from the node server instead (where it's stored.)

For this reason I've removed the api key for the purposes of this commit. I'll add it back again and continue working on the next commit.