Api call display
Again here I've removed the api key from the fetch request for the purpose of this commit.

The results now display 3 different things from the api call to the user. Everything is working fine.

I'll now work on retrieving the api key instead of hard coding it into the fetch request.