displayResults function
Here I've written a function to house the code that displays the info retrieved from the api call to the user. This cleans things up a bit in the handleSubmit function.

I've also deleted the now redundant nameChecker function. I'll continue to work on tests for the client js files.