Testing complete
After reading up on the Udacity website, the complex tax of testing the main app has been deemed unnecessary.

I have separated the grabbing of the input value into it's own function, and I've written a test for that.

I also have a test for the display results function, they both passed.

**NOTE: just in case I need it again later, I have moved the formHandler test file into the Fend-Webpack-Content repository. It has been removed from this project **